pyproject.toml
[project]
name = "web-y-tracking-supabase-jorge"
version = "0.1.0"
description = "Anti-Gravity Tracking Core for Jorge Aguirre Flores Web"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.23.0",
  "jinja2>=3.1.3",
  "python-multipart>=0.0.7",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.1.0",
  "beartype>=0.18.5",
  "orjson>=3.9.15",
  "itsdangerous>=2.1.2",
  "starlette>=0.50.0",
  "supabase>=2.3.0",
  "gotrue>=2.3.0",
  "postgrest>=2.27.3",
  "storage3>=2.27.3",
  "redis>=5.0.1",
  "upstash-redis>=1.0.0",
  "SQLAlchemy>=2.0.46",
  "psycopg2-binary>=2.9.11",
  "alembic>=1.18.3",
  "facebook-business>=18.0.0",
  "rudder-sdk-python>=2.0.2",
  "sentry-sdk[fastapi]>=1.40.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "httpx[http2]>=0.25.2",
  "structlog>=24.1.0",
  "logfire[fastapi]>=0.51.0",
  "google-auth>=2.20.0",
  "python-jose[cryptography]>=3.3.0",
  "slowapi>=0.1.9",
  "tenacity>=8.2.3",
  "phonenumbers>=8.13.0",
  "cloudflare>=2.19.0",
  "beautifulsoup4>=4.12.0",
  "typing-extensions>=4.12.0",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "locust>=2.23.1",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
  "ruff>=0.15.0",
  "mypy>=1.8.0",
  "rich>=13.7.0",
  "mock>=5.1.0",
  "playwright>=1.41.0",
  "faker>=22.6.0",
  "hypothesis>=6.92.0",
]
